+++ ui-library/src/main/java/br/com/stant/libraries/uilibrary/components/informationcardview/InformationCardViewContract.java
@@ -16,6 +16,6 @@ public interface InformationCardViewContract {
     void setDarkCardStyle();
     void setDeadlineVisibility(Integer visibility);
     void setArrowVisibility(Integer visibility);
-    void setServiceInspectionFormFilledId(Integer serviceInspectionFormFilledId);
+    void setServiceInspectionFormFilledId(Long serviceInspectionFormFilledId);
 
 }
